Add files via upload
This paper describes a new method for constructing minimal, deterministic, acyclic finite-state
automata from a set of strings. Traditional methods consist of two phases: the first to construct
a trie, the second one to minimize it. This approach is to construct a minimal automaton in a
single phase by adding new strings one by one and minimizing the resulting automaton on-the-fly.